Refactor Dockerfile for better package organization
